修改 股票执行 JOB。
