+++ jobs/cron.daily/run_daily
@@ -15,9 +15,6 @@ echo "###################"$DATETIME"###################" >> /data/logs/daily.${D
 #增加获得今日全部数据和大盘数据
 /usr/local/bin/python3 /data/stock/jobs/18h_daily_job.py >> /data/logs/daily.${DATE}.log
 
-#echo "###################"$DATETIME"###################" >> /data/logs/daily.${DATE}.log
-#增加每日，龙虎榜数据。
-#/usr/local/bin/python3 /data/stock/jobs/daily_job.py >> /data/logs/daily.${DATE}.log
 
 echo "###################"$DATETIME"###################" >> /data/logs/daily.${DATE}.log
 #使用股票指标预测。
